Add files via upload
This the weather data script. The gridMET data is all pulled from the script and no files are needed in specific directories but the historical data requires the historical weather-data file which can be found in the dropbox. Number of years of output weather data can be changed at the bottom of the script in the last section.